refactor(42-02): extract GdprDataActions from account-danger-section
- Lift 3 GDPR mutations (exportData, requestDeletion, cancelDeletion)
  + deletionStatus query + helpers (formatDeletionDate, daysRemaining)
  from account-danger-section.tsx into new gdpr-data-actions.tsx
- Introduce variant prop: 'standalone' (BlurFade + framing) | 'inline'
  (no BlurFade, tighter spacing) for reuse in subscription-cancel-section
  State 3 (upcoming task)
- account-danger-section.tsx reduced to 7-line delegate mounting
  <GdprDataActions variant="standalone" />
- No behavior change; extraction preserves exact copy + class strings
  + aria-labels from the original component

[hudsor01]
